Now we will wrap a function that is not ours with bluebird.
We'll do the mongoose connect call and we use a method of bluebird called promisify.

So first we take the method out...

and save the reference in a variable.

We call promisify on the Promise object and pass in the method we'd like to promisify and
anytime we do this, we have to pass in the object too as the second parameter.  No idea why.